V.5
Fixed background permission error. This was from when I started making
the app. Back then, extensions were the norm, so to make a packaged app,
you had to up a background permission. Somewhere along the line as
Google started pushing for actual apps, they made that the norm for
packaged apps, so  the permission became irrelevant.

Removed the button, since we are using auto-complete now.

Minor visual changes.