An intense lesson in git
So ... I lost a LOT of changes. I had made the design awesome and
responsive. I had included voice. etc. And the problem was me trying to
hurry up. I was impatient, instead of resolving errors, I hard reverted
to an old commit, and tried to copy paste over the new code from a
folder, and then pressed cntrl z ... shit happened ... and poof! Files
gone.

But, I figured if I learned anything from all that, I can redo it. And
if I didn't, I need to relearn them. Since that is the whole point of
this anyhow.